encapsulate BIP341 tapscript commitment in new SigningTarget type
I added the SigningTarget type, which encapsulates both the message
to be signed, and also the tapscript merkle root which the signature
should commit to. This allows a FROST group to dynamically select
what taproot tweak they would like to commit their signatures to,
and to optionally elect not to commit to any tweak at all if desired.